create new catalog header component
